Move animation runner to helpers
